Adds more user feature and auth system basic test

+++ src/ListingGateway.php
@@ -29,24 +29,23 @@ class ListingGateway
         return $data;
     }
 
-    public function create(array $data): string
+    public function create(int $user_id, array $data): string
     {
-        $sql = "INSERT INTO listing (name, description, price, image, author) VALUES (:name, :description, :price, :image, :author)";
+        $sql = "INSERT INTO listing (name, description, price, image, user_id) VALUES (:name, :description, :price, :image, :user_id)";
 
         $state = $this->conn->prepare($sql);
         $state->bindParam(":name", $data["name"], PDO::PARAM_STR);
         $state->bindParam(":description", $data["description"], PDO::PARAM_STR);
         $state->bindParam(":price", $data["price"], PDO::PARAM_INT);
         $state->bindParam(":image", $data["image"], PDO::PARAM_STR);
-        $mockedUserId = 0;
-        $state->bindParam(":author", $mockedUserId, PDO::PARAM_INT); // TODO : remove test mock and have proper relation
+        $state->bindParam(":user_id", $user_id, PDO::PARAM_INT);
 
         $state->execute();
 
         return $this->conn->lastInsertId();
     }
 
-    public function update(string $id, array $data): int
+    public function update(int $user_id, string $id, array $data): int
     {
         $fields = [];
 
@@ -66,8 +65,8 @@ class ListingGateway
             $fields["image"] = [$data["image"], $data["image"] === null ? PDO::PARAM_NULL : PDO::PARAM_STR];
         }
 
-        if (!empty($data["author"])) {
-            $fields["author"] = [$data["author"], PDO::PARAM_INT];
+        if (!empty($data["user_id"])) {
+            $fields["user_id"] = [$data["user_id"], PDO::PARAM_INT];
         }
 
         if (empty($fields)) {
@@ -78,10 +77,12 @@ class ListingGateway
             }, array_keys($fields));
             $sql = "UPDATE listing SET "
                 . implode(", ", $changes)
-                . " WHERE id = :id";
+                . " WHERE id = :id"
+                . " AND user_id = :user_id";
 
             $state = $this->conn->prepare($sql);
             $state->bindParam(":id", $id, PDO::PARAM_INT);
+            $state->bindParam(":user_id", $user_id, PDO::PARAM_INT);
 
             foreach ($fields as $field => $values) {
                 $state->bindValue(":$field", $values[0], $values[1]);
@@ -93,12 +94,13 @@ class ListingGateway
         }
     }
 
-    public function delete(string $id): int
+    public function delete(int $user_id, string $id): int
     {
-        $sql = "DELETE FROM listing WHERE id = :id";
+        $sql = "DELETE FROM listing WHERE id = :id AND user_id = :user_id";
 
         $state = $this->conn->prepare($sql);
         $state->bindParam(":id", $id, PDO::PARAM_INT);
+        $state->bindParam(":user_id", $user_id, PDO::PARAM_INT);
 
         $state->execute();
